requirements.txt
fastapi==0.109.1
python-multipart==0.0.9
httpx==0.27.0
duckdb==0.10.1
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.6.3
uvicorn==0.27.1